Update 6
Forgot to push the last update before making a couple more changes.  I made it so that each cycle the modulus was multiplied by 10, rather than just being set to 10 or whatever I had before.   Was having trouble with the project  not reading 7 of the numbers as true.  The lower ones were having issues with the 0's in the middle of the number and the higher ones were not being processed correctly because of data limits.  I made some adjustments to the "if" statement to make up for it, as im not sure how else I would have solved this.